Leave union overlay method intact for netcore
